Changed the way 'MonteCarlo' simulations work for the stochastic network model (#21)
* Updated plotting in fcn plotPopulationStatus

* Updated plotting in fcn plotFit

* Finished updating all plotting functions with Daans Style

* Added PSO solver in function fit of SEIRSNetworkModel

All genetic algorithms will be replaced by PSO later on. The change is necessary to run an optimisation on HPC.

* Added a script to calibrate stochastic model on HPC

* changed the way 'Monte-Carlo' sampling works in stochastic model

Each node in network is assigned an incubation time drawn from the probability distribution. This is enabled by setting monteCarlo to True. The  variable n_samples was removed. Rather, the user can choose to perform the simulation multiple times by defining a number of 'repeats'.