Tryout spatial model (#149)
This is a first technical implementation of a spatial modeling framework to monitor COVID-19 spread at the level of the Belgian arrondissements. It builds upon the stochastic national-level model to take spatial heterogeneity into account. The Belgian territory is divided into 43 arrondissements, which are from hereon referred to as patches. Our patch-model boils down to a simulation of the extended SEIRD model dynamics on each patch, where the patches are connected by commuting traffic (individuals going to work). This 'connection' is a 43x43 from-to commuting matrix, named `space`, extracted from the 2011 Belgian census data (see data readme).