Update notebook working environment for easy use
Currently the show_maps and show_graphs functions do not function because of an error related to the `quantile` function of xarray:
``
ValueError: dimension 'draws' on 0th function argument to apply_ufunc with dask='parallelized' consists of multiple chunks, but is also a core dimension. To fix, rechunk into a single dask array chunk along this dimension, i.e., ``.chunk({'draws': -1})``, but beware that this may significantly increase memory usage.
``