Bugfixing workflow test (#229)
The MCMC sampler in the workflow test test_workflow_COVID19_SEIRD.py would block if a low number of PSO iterations were performed. The starting positions for the sampler would not be sufficiently independent of eachother. I surpassed this by running a long PSO run and hardcode the obtained parameter estimates.